Working web server!
The server is still single threaded and doesn't use event polling, so
the scalability is severly limited. That will be addressed in a future
revision.

The file handler class provided in the test application is intended to
be very basic. There is no configuration and it does not understand how
to follow symbolic links. However, this is intended only as an example
of how to write a handler and insert it into the router.