Added file handler to project.
After working with the file handler in the demo server application, it
made sense to make it a permanent part of the project. It now depends on
the `mapped` submodule to read the file and handle errors.

Also the HTTP method was changed from a string to a bit flag/enumeration
so that checks can be more easily performed in the router and not in the
actual handler.